Improve categorical compare and add utils
