+++ src/transformations/util.ts
@@ -1,6 +1,7 @@
 import { Collection, CodapAttribute } from "../utils/codapPhone/types";
 import { Env } from "../language/interpret";
 import { Value } from "../language/ast";
+import { DataSet } from "./types";
 
 /**
  * Converts a data item object into an environment for our language. Only
@@ -142,68 +143,21 @@ export function uniqueAttrName(base: string, attrs: CodapAttribute[]): string {
   return name;
 }
 
-/**
- * Compute the union of two arrays, using `pred` to determine equality
- */
-export function unionWithPredicate<T>(
-  array1: T[],
-  array2: T[],
-  pred: (elt1: T, elt2: T) => boolean
-): T[] {
-  const out: T[] = [];
-
-  const merged: T[] = array1.concat(array2);
-
-  for (const elementToBeAdded of merged) {
-    if (
-      out.find((existingElement: T) =>
-        pred(existingElement, elementToBeAdded)
-      ) === undefined
-    ) {
-      out.push(elementToBeAdded);
-    }
+export function getAttributeDataFromDataset(
+  attributeName: string,
+  dataset: DataSet
+): CodapAttribute {
+  let attributeData: CodapAttribute | undefined;
+  for (const collection of dataset.collections) {
+    attributeData =
+      collection.attrs?.find((attribute) => attribute.name === attributeName) ??
+      attributeData;
+  }
+  if (!attributeData) {
+    throw new Error(
+      "Couldn't find first selected attribute in selected context"
+    );
   }
 
-  return out;
-}
-
-/**
- * Compute the intersection of two arrays, using `pred` to determine equality
- */
-export function intersectionWithPredicate<T>(
-  array1: T[],
-  array2: T[],
-  pred: (elt1: T, elt2: T) => boolean
-): T[] {
-  return array1.filter(
-    (elt1) => array2.find((elt2) => pred(elt1, elt2)) !== undefined
-  );
-}
-
-/**
- * Compute the difference of two arrays, using `pred` to determine equality
- */
-export function setDifferenceWithPredicate<T>(
-  array1: T[],
-  array2: T[],
-  pred: (elt1: T, elt2: T) => boolean
-): T[] {
-  return array1.filter(
-    (elt1) => array2.find((elt2) => pred(elt1, elt2)) === undefined
-  );
-}
-
-/**
- * Compute the symmetric difference of two arrays, using `pred` to determine equality
- */
-export function symmetricDifferenceWithPredicate<T>(
-  array1: T[],
-  array2: T[],
-  pred: (elt1: T, elt2: T) => boolean
-): T[] {
-  return unionWithPredicate(
-    setDifferenceWithPredicate(array1, array2, pred),
-    setDifferenceWithPredicate(array2, array1, pred),
-    pred
-  );
+  return attributeData;
 }
